refactor: Price 도메인 common 패키지로 이동
