+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,16 +1,12 @@
 package kitchenpos.product.domain;
 
-import static kitchenpos.exception.ErrorCode.PRICE_NOT_EXISTS_OR_LESS_THAN_ZERO;
-
 import java.math.BigDecimal;
-import java.util.Objects;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
-import kitchenpos.exception.KitchenposException;
+import kitchenpos.common.domain.Price;
 
 @Entity
 public class Product {
